I got the ned to ecef converstion working.  The velocity calculation is still not good enough.  It flies to waypoints, but it is way underdamped.  If I use true velocity that fixes this problem.
