+++ roscopter_sim/launch/mocap_landing.launch
@@ -53,7 +53,8 @@
 
 
   <group ns="m2u"> <!--m2u = motion capture to ublox conversion-->
-    <node name="mocap2ublox" pkg="roscopter" type="mocap2ublox_ros.py" output="screen">
+    <node name="mocap2ublox" pkg="roscopter" type="mocap2ublox2.py" output="screen">
+      <remap from="gnss" to="/$(arg mav_name)/gps/data"/>
       <remap from="rover_mocap" to="/sim/rover_mocap"/>
       <remap from="base_mocap" to="/sim/base_mocap"/>
     </node>
@@ -109,8 +110,8 @@
       <!-- <remap from="reference" to="ground_truth/odometry/NED_throttled"/> -->
       <!-- <remap from="gnss" to="gps/data"/> -->
       <remap from="ublox_gnss" to="/m2u/rover_PosVelEcef"/>
-      <remap from="ublox_relpos" to="/m2u/rover_relpos"/>
-      <remap from="ublox_base_posvelecef" to="/m2u/base_PosVelEcef"/>
+      <!-- <remap from="ublox_relpos" to="/m2u/rover_relpos"/> -->
+      <!-- <remap from="ublox_base_posvelecef" to="/m2u/base_PosVelEcef"/> -->
       <!--Outputs-->
       <remap from="is_flying" to="not_for_sim"/>
     </node>
